Added "file stack line accumulator"
I ended up going with my program of 2017. I boiled away the nastier parts... but there is still some amount of weirdness going on. Said weirdness stands on a grand total of 4 lines... which I do not value for the hours it will take the reorganize the code to remove